Fix metrics value type (#284)
* Starting draft (use in educational purposes only)

* ddos-guard example

* move ddos-guard example into advanced folder

* change metrics value types to uint64

+++ src/jobs/packetgen.go
@@ -100,7 +100,7 @@ func packetgenJob(ctx context.Context, globalConfig GlobalConfig, args Args, deb
 			protocolLabelValue,
 			metrics.StatusSuccess)
 
-		trafficMonitor.Add(len)
+		trafficMonitor.Add(uint64(len))
 	}
 
 	return nil, nil
